pyproject.toml
[tool.poetry]
name = "try_pipelining"
version = "0.1.0"
description = "little test project to explore pipelining of tasks and filtering"
authors = ["Clemens Hoischen <clemenshoischen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
astropy = "^4.3.1"
ephem = "^4.1"
matplotlib = "^3.4.3"
pydantic = "^1.8.2"
pytz = "^2021.3"
rich = "^10.12.0"
isort = "^5.10.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
